Change the handling of non-responsive commands against the gateway.
Since not all commands return a response, we had the executor set to timeout after 1 second and move on to the next command. This worked great on our small scale system, but apparently larger scale systems do not responsd as quickly when querying loads. This changes the timeout to 8 seconds instead.

The timeout is almost never used in our system anyways. Since we found out that we can string together commands, instead of executing them one-by-one, any buttons or zones that you turn on should really always return a response, since we know the status of the loads beforehand and only change to valid statuses.