only query the active loads once when app is initially created.
Since it was querying the loads from the onResume method, which runs after the onCreate method, it was forcing the app to attempt to query the active loads twice whenever you initially opened the app. This seemed to cause connection problems with the socket.

Connection problems when querying the loads seem to be resolved after ensuring they are only queried a single time.